# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Todo List API
  version: 1.1.0
  description: REST API for authentication and todo management.
servers:
  - url: http://localhost:3000
tags:
  - name: System
  - name: Auth
  - name: Todos
paths:
  /:
    get:
      tags:
        - System
      summary: Health check
      responses:
        "200":
          description: API is running
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
  /api/auth/register:
    post:
      tags:
        - Auth
      summary: Register a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "201":
          description: Registered user and access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/auth/login:
    post:
      tags:
        - Auth
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "200":
          description: Authenticated user and access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/auth/me:
    get:
      tags:
        - Auth
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/todos:
    get:
      tags:
        - Todos
      summary: List todos
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Todo list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Todo"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - Todos
      summary: Create a todo
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTodoRequest"
      responses:
        "201":
          description: Created todo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/todos/{id}:
    patch:
      tags:
        - Todos
      summary: Mark a todo as completed
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TodoID"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags:
        - Todos
      summary: Delete a todo
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TodoID"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TodoID:
      name: id
      in: path
      required: true
      description: MongoDB ObjectID of the todo.
      schema:
        type: string
        example: 6636d3d046b1b2dd3b46e001
  responses:
    Success:
      description: Operation succeeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessResponse"
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: 6636d3d046b1b2dd3b46e001
        email:
          type: string
          format: email
          example: user@example.com
        createdAt:
          type: string
          format: date-time
    AuthRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 6
          example: secret123
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
          description: Bearer token for authenticated requests.
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
    Todo:
      type: object
      properties:
        id:
          type: string
          example: 6636d3d046b1b2dd3b46e001
        userId:
          type: string
          example: 6636d3d046b1b2dd3b46e002
        completed:
          type: boolean
          example: false
        body:
          type: string
          example: Learn hexagonal architecture
    CreateTodoRequest:
      type: object
      required:
        - body
      properties:
        body:
          type: string
          example: Learn Fiber
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: invalid credentials